Fix broken download recipe
The make variables that expand to commands will have a semicolon at the
end already.  Then one after it puts two in a row, which isn't valid
shell.  It's ok in C, but in a shell script, there needs to be a command
betweeen semicolons.

Then the @echo will also be invalid because the @ needs to at the start
of a line, since it's not part of the command, but a signal to make, and
it's only recognized at the start of a line.

This whole case statement here with the linux tarball check looks like a
total kludge and should die.