gcc: Turn off sanitizer with uClibc
From what I can tell, this just doesn't work at all with uClibc.  It
probes into libc internals that don't work on uClibc.  The sanitizer
feature is somewhat new and these toolchain makefiles were probably
copied from openwrt before it existed.

The sanitizer is a debugging aid, which won't even be enabled or used on
the R7800 firmware build, so no reason to have it.