Build with unpatched toolchains
The toolchain files are from an old openwrt release and don't support
new toolchains.

The default config is to use gcc 10.2.0, which is too new to have
openwrt patches here.  Might be some in current openwrt, which could be
pulled in, but I haven't done that.

So the toolchain is unpatched.  One of those missing patches adds
support for a special compiler flag, which makes it so nothing builds
without the patch.

Attempt to detect patched toolchains, and only use this flag with them,
so that unpatched toolchains like 10.2.0 still work.