Add package for the Wifi FW
This is refenced in the deconfig, but no package exists for it.

Create one that will just install a binary of the firmware in the right
place.  There's a reference to a tarball that has it, but that file
can't be found anywhere.

The wifi fw files will need to be extracted from an R7800 firmware
image.